# Port System.Memory performance improvements to Convert.TryFromBase64Chars