Fixes https://github.com/dotnet/corefx/issues/24897