@jkotas 2018-03-19
@dotnet-bot test Windows_NT x86 Checked Innerloop Build and Test please