@jkotas 2018-03-19
Some data to verify that it is really improving performance compared to what we have there today may be useful.