@AtsushiKan 2018-03-20
Some quick perf numbers:

All of these done with the input being a 500 byte array that was converted to Base64 (so ~667 characters.) and iterating 1000000 times:

- No whitespace: 
      old: 2,110.79 ms
      new: 525.08 ms
    
    4x improvement.

- Newline every 76 characters (76 or 64 is the standard: Base64FormattingOptions.InsertLines gives you 76)

      old: 2,110.79 ms
      new: 525.08 ms
    
    4x improvement.

      old: 2,230.02 ms
      new: 675.57 ms

    3.4x improvement.

The tradeoff is if line breaks ocurr within 4-byte chunks: shouldn't be the typical case but we can take a perf hit there. As an extreme case:

    Space inserted every 3 characters:

      old: 3,029.01 ms
      new: 5,874.91 ms

     1.9x regression.

The penalty does naturally shrink (or disappear) as the whitespace percentage goes down.

So that's the pricetag for not keeping the old algorithm around as a fallback for this weird case.