@stephentoub 2018-03-20
Yeah, that's excellent, thanks for persisting in bringing these improvements over.