@AtsushiKan 2018-03-20
@dotnet-bot test Tizen armel Cross Checked Innerloop Build and Test
@dotnet-bot test Ubuntu x64 Checked Build and Test (Jit - CoreFx)