@AtsushiKan 2018-03-20
@dotnet-bot test Ubuntu x64 Checked Build and Test